Get the latest - 2017/10/18

+++ framework/render/gles30_v2/ogles_3_0_renderer_v2.cpp
@@ -309,7 +309,7 @@ class Spritesheet
 		while (!validEntries.empty())
 		{
 			LogInfo("Repack: creating sheet %d", (int)pages.size());
-			auto page = mksp<SpritesheetPage>(pages.size(), page_size, node_count);
+			auto page = mksp<SpritesheetPage>((int)pages.size(), page_size, node_count);
 			pages.push_back(page);
 			page->addMultiple(validEntries);
 
@@ -341,7 +341,7 @@ class Spritesheet
 		}
 		// Required a new page
 		LogInfo("Creating spritesheet page %d", (int)pages.size());
-		auto page = mksp<SpritesheetPage>(pages.size(), page_size, node_count);
+		auto page = mksp<SpritesheetPage>((int)pages.size(), page_size, node_count);
 		auto ret = page->addEntry(entry);
 		if (!ret)
 		{
